Adding URL Params, Auto Superscripts, + Easter Egg
Added a URL Parameter that forces itself onto every link on the page.
The URL param is set up so that when it's used the content on that page
becomes ungated, so by having it push to every link on the page, it
allows for a version of the site that is essentially ungated (Currentlly
calling it the VIP Version).

Added jquery to automatically superscript any trademarks.

Hid the slide in "Next Flipbook" while on the ungated version of the
site because they are still gating currently.

Created a referral param that ungates a single page on the site.

Added a small easter egg that will be deleted :smile: .

Simiplified and clarified the CSS and HTML.

Edited Alt Tags for the image map and changed from href to a href to the
url param passing works on the image map.