Resolving Issue #3  with URL Params, New Meta Tags
Resolved issue #3, thought the problem was with the images loading
twice, but it was a bug with the code that forwards the URL parameter.
I found a better solution online for forwarding URL params and used that
instead.

Added the old "iguscta" param incase anyone was still using those links.

Worked with uberflip to find a workaround to get the <meta> tags in the
header since it's not normally accessible through their platform.

Changed so header shows with all url parameters on the homepage.